requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Elasticsearch 
elasticsearch>=8.0.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Data processing
python-dateutil>=2.8.0

# Utilities
tenacity>=8.2.0

